'Refactored by Sourcery'

+++ scripts/calculate_ci_perf.py
@@ -86,12 +86,11 @@ def _github_repo(repo_url: str, sha: Optional[str], repo_destination: Path) -> P
         repo_sha = subprocess.check_output(
             ["git", "rev-parse", "HEAD"], stderr=subprocess.STDOUT
         )
-        if sha:
-            if not repo_sha.startswith(sha.encode("utf-8")):
-                shutil.rmtree(repo_destination)
-                raise GitError(
-                    f"Github repo is broken (not set to correct sha: {repo_sha.decode('utf-8')}"
-                )
+        if sha and not repo_sha.startswith(sha.encode("utf-8")):
+            shutil.rmtree(repo_destination)
+            raise GitError(
+                f"Github repo is broken (not set to correct sha: {repo_sha.decode('utf-8')}"
+            )
 
     return repo_destination
